feat: day 6 part 2
probably the worst code i've ever written. i tried really hard to find
the loop by matching with a saved position and direction, but it only
led to infinite loops that weren't detected, so i thought "hell, why
not" and just made a safety net to catch loops which turn counters
exceeded the allowed max turns amount. i'm ashamed of this code,
honestly